scroll-animations/scroll-timeline-invalidation.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollTimeline invalidation</title>
<link rel="help" href="https://wicg.github.io/scroll-animations/#current-time-algorithm">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<style>
  .scroller {
    overflow: auto;
    height: 100px;
    width: 100px;
  }
  .contents {
    height: 1000px;
    width: 100%;
  }
</style>
<div id="log"></div>

<script>
'use strict';
function createScroller(test) {
  var scroller = createDiv(test);
  scroller.innerHTML = "<div class='contents' id='contents'></div>";
  scroller.classList.add('scroller');
  return scroller;
}

function createScrollTimeline(test) {
  return new ScrollTimeline({
    scrollSource: createScroller(test),
    timeRange: 1000
  });
}

function createScrollLinkedAnimation(test, timeline) {
  if(timeline === undefined)
    timeline = createScrollTimeline(test);
  const DURATION = 350; // ms
  const KEYFRAMES = { opacity: [1, 0] };
  return new Animation(
    new KeyframeEffect(createDiv(test), KEYFRAMES, DURATION), timeline);
}

promise_test(async t => {
  const animation = createScrollLinkedAnimation(t);
  const scroller = animation.timeline.scrollSource;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;

  animation.play();
  await animation.ready;

  scroller.scrollTop = 0.2 * maxScroll;

  // Change scroller content size.
  document.getElementById("contents").style.height = "500px";

  await animation.finished;
  const newTime = animation.effect.getTiming().duration;
  assert_times_equal(animation.currentTime, newTime,
    'Animation current time is updated after scroller invalidation.');

  assert_times_equal(
    animation.effect.getComputedTiming().localTime, newTime,
    'Effect local time is updated after scroller invalidation.');
}, 'Animation current time and effect local time are updated after scroller ' +
   'content size changes.');

promise_test(async t => {
  const animation = createScrollLinkedAnimation(t);
  const scroller = animation.timeline.scrollSource;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;

  animation.play();
  await animation.ready;

  scroller.scrollTop = 0.2 * maxScroll;

  // Change scroller size.
  scroller.style.height = "500px";

  await animation.finished;
  const newTime = animation.effect.getTiming().duration;
  assert_times_equal(animation.currentTime, newTime,
    'Animation current time is updated after scroller invalidation.');

  assert_times_equal(
    animation.effect.getComputedTiming().localTime, newTime,
    'Effect local time is updated after scroller invalidation.');
}, 'Animation current time and effect local time are updated after scroller ' +
   'size changes.');

</script>
